Move shared files under tools/Common/MachO
