+++ src/coreclr/tools/Common/MachO/BinaryFormat/LoadCommand.cs
@@ -3,7 +3,11 @@
 
 using System.Runtime.InteropServices;
 
+#if HOST_MODEL
 namespace Microsoft.NET.HostModel.MachO;
+#else
+namespace ILCompiler.Reflection.ReadyToRun.MachO;
+#endif
 
 /// <summary>
 /// The base structure for all load commands in a Mach-O binary.
