+++ src/coreclr/tools/Common/MachO/BinaryFormat/MachHeader.cs
@@ -3,7 +3,11 @@
 
 using System.Runtime.InteropServices;
 
+#if HOST_MODEL
 namespace Microsoft.NET.HostModel.MachO;
+#else
+namespace ILCompiler.Reflection.ReadyToRun.MachO;
+#endif
 
 /// <summary>
 /// The Mach-O header is the first data in a Mach-O file.
