+++ src/coreclr/tools/Common/MachO/BinaryFormat/Section64LoadCommand.cs
@@ -3,7 +3,11 @@
 
 using System.Runtime.InteropServices;
 
+#if HOST_MODEL
 namespace Microsoft.NET.HostModel.MachO;
+#else
+namespace ILCompiler.Reflection.ReadyToRun.MachO;
+#endif
 
 /// <summary>
 /// A load command that provides information about a section in a segment.
