+++ src/coreclr/tools/aot/ILCompiler.Reflection.ReadyToRun/ILCompiler.Reflection.ReadyToRun.csproj
@@ -31,18 +31,8 @@
     <Compile Include="..\..\Common\Internal\Runtime\ModuleHeaders.cs" Link="Common\ModuleHeaders.cs" />
     <Compile Include="..\..\Common\Internal\Runtime\ReadyToRunConstants.cs" Link="Common\ReadyToRunConstants.cs" />
     <Compile Include="..\..\Common\Internal\Runtime\ReadyToRunInstructionSet.cs" Link="Common\ReadyToRunInstructionSet.cs" />
+    <Compile Include="..\..\Common\MachO\**\*.cs" Link="MachO\%(RecursiveDir)%(Filename)%(Extension)" />
     <Compile Include="..\..\Common\Pgo\PgoFormat.cs" Link="Common\PgoFormat.cs" />
   </ItemGroup>
 
-  <ItemGroup>
-    <!-- Include MachO files from Microsoft.NET.HostModel -->
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\MachO\*.cs" Link="MachO\%(Filename)%(Extension)" />
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\MachO\Enums\*.cs" Link="MachO\Enums\%(Filename)%(Extension)" />
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\MachO\BinaryFormat\*.cs" Link="MachO\BinaryFormat\%(Filename)%(Extension)" />
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\MachO\BinaryFormat\Blobs\*.cs" Link="MachO\BinaryFormat\Blobs\%(Filename)%(Extension)" />
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\MachO\FileUtils\*.cs" Link="MachO\FileUtils\%(Filename)%(Extension)" />
-
-    <!-- Remove -->
-    <Compile Include="..\..\..\..\installer\managed\Microsoft.NET.HostModel\AppHost\AppHostExceptions.cs" Link="AppHost\%(Filename)%(Extension)" />
-  </ItemGroup>
 </Project>
